Created the hello_world.py file
I was able to add this file directly through the GitHub website. I was not able to get git to work through VSCode for some reason. It's been especially painful over the last few days and keeps saying that there are no changes to commit even when there definitely are...